Refactor command classes into command package
