+++ src/main/java/lania/Parser.java
@@ -1,5 +1,12 @@
 package lania;
 
+import lania.command.AddCommand;
+import lania.command.Command;
+import lania.command.CompleteCommand;
+import lania.command.DeleteCommand;
+import lania.command.ExitCommand;
+import lania.command.FindCommand;
+import lania.command.ListCommand;
 import lania.exception.LaniaEmptyDescriptionException;
 import lania.exception.LaniaException;
 import lania.task.Deadline;
